salesforce_list_targeting/requirements.txt
beatbox-0.96.zip
PyYAML==3.11
requests==2.7.0
